AirPumpPir - Control your air pump or air conditioner with an arduino, PIR motion sensors and DHT22 temperature sensor
An air pump and air conditioner are most efficient when at high fan
speed.  I was however disturbed by the noise this gives so when I enter
the room I would like the fan speed to decrease and after I leave the
room the fan speed could go high again. It is a two step process, first
you read the data from you remote and save the raw signal. You then
enter this in the other code for run time.